+++ CHANGELOG.md
@@ -11,10 +11,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - added ensure_adminuser command to add subtly more sophisticated behaviour to Django's createsuperuser - i.e. don't raise non-zero exit code if superuser already exists, add more useful output.
 - Use uv in the docker image.
+- Empty arrays to empty responses for a consistent response structure.
+
+### Changed
+- Inappropriate use of 204 code: RecentRoutesView changed from 204 to 200 OK with an empty array and the original message ("No recent routes found for today.").
+- Inappropriate use of 204 code: VehicleTypeListView changed from 204 to 200 OK with and empty array and the original message.
+- MeshView - Changed from 204 to 404 Not Found when mesh doesn't exist.
+- Updated tests to reflect corrected HTTP status codes.
 
 ### Fixed
 - Add erroneously missing `rest_framework` into `INSTALLED_APPS`.
 - Remove unique constraint and add id field to locations fixture to prevent duplication.
+
+
+
 
 
 ## 0.2.4 - 2025-11-11